force no cache avatar. Refacto Avatar. Cleaning old Settings. Cleaning Profile
